refactor table column

+++ database/migrations/2022_11_19_120359_create_store_table.php
@@ -17,7 +17,7 @@ class CreateStoreTable extends Migration
             $table->integer('id', true);
             $table->string('name', 45)->nullable();
             $table->string('location', 45)->nullable();
-            $table->integer('rating')->nullable();
+            $table->double('rating')->nullable();
         });
     }
